# tracker/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "tracker",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:3003"
      }
    ],
    "paths": {
      "/tracker/messages": {
        "get": {
          "summary": "Tracker retrieve message by text",
          "tags": [],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "description": "",
                    "minItems": 1,
                    "uniqueItems": true,
                    "items": {
                      "type": "object",
                      "required": [],
                      "properties": {}
                    },
                    "x-examples": {
                      "example-1": [
                        "good morning",
                        "good night"
                      ]
                    }
                  }
                }
              }
            }
          },
          "operationId": "get-tracker-messages",
          "parameters": [
            {
              "schema": {
                "type": "string"
              },
              "in": "header",
              "name": "Authorization",
              "required": true
            },
            {
              "schema": {
                "type": "string"
              },
              "in": "query",
              "name": "text"
            }
          ]
        }
      },
      "/tracker/count_messages": {
        "get": {
          "summary": "Tracker retrieve message count",
          "tags": [],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "description": "",
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "number"
                      },
                      "ok": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "message",
                      "ok"
                    ],
                    "x-examples": {
                      "example-1": {
                        "message": 3,
                        "ok": true
                      }
                    }
                  }
                }
              }
            }
          },
          "operationId": "get-tracker-count_messages",
          "parameters": [
            {
              "schema": {
                "type": "string"
              },
              "in": "header",
              "name": "Authorization"
            },
            {
              "schema": {
                "type": "string"
              },
              "in": "query",
              "name": "category",
              "required": true
            },
            {
              "schema": {
                "type": "string"
              },
              "in": "query",
              "name": "created_time",
              "required": true
            }
          ]
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "title": "User",
          "type": "object",
          "description": "",
          "examples": [
            {
              "id": 142,
              "firstName": "Alice",
              "lastName": "Smith",
              "email": "alice.smith@gmail.com",
              "dateOfBirth": "1997-10-31",
              "emailVerified": true,
              "signUpDate": "2019-08-24"
            }
          ],
          "properties": {
            "id": {
              "type": "integer",
              "description": "Unique identifier for the given user."
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string",
              "format": "email"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date",
              "example": "1997-10-31"
            },
            "emailVerified": {
              "type": "boolean",
              "description": "Set to true if the user's email has been verified."
            },
            "createDate": {
              "type": "string",
              "format": "date",
              "description": "The date that the user was created."
            }
          },
          "required": [
            "id",
            "firstName",
            "lastName",
            "email",
            "emailVerified"
          ]
        }
      }
    }
  }